docs: Update LoaderAndLayerInterface doc
Clean-up and update the LoaderAndLayerInterface document so that
it is easier to navigate and follow.  Add in a Table-of-Contents
at the beginning, and navigation links throughout.  Also, re-arranged
the sections so that all the Loader-Layer interface discussion is under
one primary section, and the Loader-ICD interface discussions are under
another section.  Previously, there were several sections for each.

Special thanks to Lenny who contributed new images, and several
rewritten sections.

Change-Id: I5cb4af31171f8bf56374d09ad0e95642cf8b0fa8