WIP: Added migrate feature to spec.py script
Added '-migrate' option to spec.py.
This is a temp option that will read in a spec that's been updated
to include the error ids in-line, and then migrate from the existing
database file to use the ids from the new spec. This will also go
through the source files and replace old ids with new ids.

The only definite hole in this is that the spec parser needs to be
updated to handle the new valid usage format.
If the unique id is in-line at the start of valid usage text, then need
to kill that portion of string as well.