layers:Improve vtx buffer perf warning
Improve perf warning related to vtx buffers being bound to a cmd buffer
but not being used. Track when vtx buffers are used, and then only flag
the perf warning if no vtx buffer has ever been used at the time when
a draw occurs with a pipeline that has vtx binding state.